+++ server/src/api/analytics/runDashboardCardQuery.ts
@@ -99,7 +99,7 @@ export async function runDashboardCardQuery(
     WITH scoped_events AS (
       SELECT *
       FROM events
-      PREWHERE site_id IN {siteIds:Array(UInt16)}
+      PREWHERE site_id IN {siteIds:Array(UInt32)}
       WHERE 1=1 ${timeStatement}
     )
     SELECT *
